+++ tests/unit/Docs/SongStatisticRequestTest.php
@@ -1,0 +1,75 @@
+<?php
+
+
+namespace Tests\Unit\Docs;
+
+
+use CTApi\Requests\SongRequest;
+use CTApi\Requests\SongStatisticRequest;
+use CTApi\Requests\SongStatisticRequestBuilder;
+use Tests\Unit\TestCaseHttpMocked;
+
+class SongStatisticRequestTest extends TestCaseHttpMocked
+{
+
+    public function testGetAll()
+    {
+        $data = SongStatisticRequest::all();
+
+        $songStatistic = end($data);
+
+        $this->assertEquals(14, $songStatistic->getCount());
+        $this->assertEquals(11, $songStatistic->getCountForCalendars([1, 2]));
+
+        $song = $songStatistic->requestSong();
+
+        // Retrieve Dates:
+        $allDates = $songStatistic->getDates();
+        $date = end($allDates);
+        $this->assertEquals('2021-12-12 10:30:00', $date["date"]);
+        $this->assertEquals('3', $date["calendar_id"]);
+
+        $datesForMyServies = $songStatistic->getDatesForCalendars([1, 2]);
+        $dateService = end($datesForMyServies);
+        $this->assertEquals('2021-07-11 10:30:00', $dateService["date"]);
+        $this->assertEquals('2', $dateService["calendar_id"]);
+    }
+
+    public function testGetViaSong()
+    {
+        $song = SongRequest::findOrFail(21);
+        $statistics = $song->requestSongStatistic();
+
+        $this->assertEquals($statistics?->getCount(), 8);
+        $this->assertEquals($statistics?->getCountForCalendars([2]), 5);
+        $this->assertEquals($statistics?->getCountForCalendars([21]), 0);
+    }
+
+    /**
+     * @doesNotPerformAssertions
+     */
+    public function testLazy()
+    {
+        /**
+         * Lazy-SongStatisticRequestBuilder
+         */
+        $requestBuilderLazy = new SongStatisticRequestBuilder(); // default: lazy-flag is true
+
+        $requestBuilderLazy->find("21");
+        $requestBuilderLazy->find("22");
+
+        // The whole song-statistic will be fetched ones and used for both "find"-calls.
+
+        /**
+         * Not-Lazy-SongStatisticRequestBuilder
+         */
+
+        $requestBuilderNotLazy = new SongStatisticRequestBuilder(false);
+
+        $requestBuilderLazy->find("21");
+        $requestBuilderLazy->find("22");
+
+        // The whole song-statistic will be fetched for every "find"-call.
+    }
+
+}